Symmetric procrustes
We use `symmetric = TRUE` when calling procrustes. That gives a
symmetric procrustes statistic, and a symmetric global SS, i.e. the
order of the distance matrices is not important (plant/pollinator =
pollinator/pant). However, because the PCO matrix of each jackknife
link is different, the procrustes normalisation is also different. That
causes the paco-links weights to be non-symmetric.